+++ verik-compiler/src/main/kotlin/io/verik/compiler/serialize/source/ExpressionSerializer.kt
@@ -31,7 +31,6 @@ import io.verik.compiler.ast.element.expression.sv.EArrayLiteralExpression
 import io.verik.compiler.ast.element.expression.sv.ECaseStatement
 import io.verik.compiler.ast.element.expression.sv.EConcatenationExpression
 import io.verik.compiler.ast.element.expression.sv.EConstantPartSelectExpression
-import io.verik.compiler.ast.element.expression.sv.EDelayExpression
 import io.verik.compiler.ast.element.expression.sv.EEventControlExpression
 import io.verik.compiler.ast.element.expression.sv.EEventExpression
 import io.verik.compiler.ast.element.expression.sv.EForeverStatement
@@ -184,6 +183,10 @@ object ExpressionSerializer {
 
     fun serializeSuperExpression(serializeContext: SerializeContext) {
         serializeContext.append("super")
+    }
+
+    fun serializeBreakStatement(serializeContext: SerializeContext) {
+        serializeContext.appendLine("break;")
     }
 
     fun serializeReturnStatement(returnStatement: EReturnStatement, serializeContext: SerializeContext) {
@@ -417,11 +420,6 @@ object ExpressionSerializer {
             serializeContext.serializeAsExpression(it)
         }
         serializeContext.append(")")
-    }
-
-    fun serializeDelayExpression(delayExpression: EDelayExpression, serializeContext: SerializeContext) {
-        serializeContext.append("#")
-        serializeContext.serializeAsExpression(delayExpression.expression)
     }
 
     private fun serializePropertyInline(property: EProperty, serializeContext: SerializeContext) {